feat: 카테고리 리팩토링
