+++ package.json
@@ -12,7 +12,7 @@
 		"axios": "^0.27.2",
 		"global": "^4.4.0",
 		"react": "18.2.0",
-		"react-carousel-minimal": "^1.4.1",
+		"react-carousel-minimal": "1.4.1",
 		"react-cookie": "4.1.1",
 		"react-daum-postcode": "^3.1.1",
 		"react-dom": "^18.2.0",
