+++ src/pages/Shop.jsx
@@ -1,4 +1,4 @@
-import Category from 'components/shop/category';
+import Category from 'components/shop/Category';
 import { useNavigate, useParams } from 'react-router-dom';
 import { useEffect, useState } from 'react';
 import axios from 'axios';
